Move the `__rt` module into its own file (#4343)
